Indicate initial schema version is `10.2.0` (instead of `10.3.0`)
This makes it consistent with the work-in-progress migration notebook.
